Fix request with external service

+++ pagamento-service/src/main/java/com/fiap/pagamentoservice/infrastructure/external/dto/ProcpagRequest.java
@@ -1,9 +1,7 @@
 package com.fiap.pagamentoservice.infrastructure.external.dto;
 
-import java.math.BigDecimal;
-
 public record ProcpagRequest(
-        BigDecimal valor,
+        Integer valor,
         String pagamento_id,
         String cliente_id
 ) {
